competition-test/requirements.txt
alpaca-py
pandas
numpy
python-dotenv
pytz
torch
